% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringmanipulation.R
\name{birthdayFromCPR}
\alias{birthdayFromCPR}
\title{A function for calculating the birthday from a CPR number, taking the century into account}
\usage{
birthdayFromCPR(cpr)
}
\arguments{
\item{cpr}{CPR number. Any format will work.}
}
\value{
a Date of the birthday
}
\description{
A function for calculating the birthday from a CPR number, taking the century into account
}
\examples{

birthdayFromCPR("290511-4487")
returns "2011-05-29"

birthdayFromCPR("290511-3487")
returns "1911-05-29"


}
