% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectormanipulation.R
\name{combineVectors}
\alias{combineVectors}
\title{A function for collapsing 2 or more vectors containing NA values
Requires all vectors to have the same length}
\usage{
combineVectors(...)
}
\arguments{
\item{...}{n vectors}
}
\value{
a single vector
}
\description{
Collapses from first to last
}
\examples{
v1 = c(1,NA,NA)
v2 = c(2,3,NA)
v3 = c(4,5,6)

combineVectors(v1, v2, v3)
#returns 1, 3, 6

}
