% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{errorcare}
\alias{errorcare}
\title{A wrapper for trycatch, which allows ignoreing of errors and leave the loop running.
This is terrible behavior and should only be used when necessary}
\usage{
errorcare(...)
}
\arguments{
\item{...}{whatever function you want to ignore errors from}
}
\value{
output of the function, or NA if something failed
}
\description{
A wrapper for trycatch, which allows ignoreing of errors and leave the loop running.
This is terrible behavior and should only be used when necessary
}
\examples{
data = matrix(rnorm(40), ncol = 8)
data[,5] = NA

for(i in 2:ncol(data)) {
  result = t.test(data[,(i-1)], data[,i])$p.value
  print(result)
}
#This halts execution at the 5'th cloumn

for(i in 2:ncol(data)) {
  result = t.test(data[,(i-1)], data[,i])$p.value
  print(result)
}
#Works, but returns NA when column 5 is being used

}
