% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggkm}
\alias{ggkm}
\title{A function for Kaplan-Meier plots with ggplot2}
\usage{
ggkm(sfit, table = TRUE, returns = FALSE, xlabs = "Time",
  ylabs = "Survival Probability", xlims = c(0, max(sfit$time)),
  ylims = c(0, 1), ystratalabs = NULL, ystrataname = NULL,
  timeby = 100, main = "Kaplan-Meier Plot", pval = TRUE,
  subs = NULL, linetypes = NULL, ...)
}
\arguments{
\item{sfit}{A survfit object from Survival package}
}
\description{
Found at https://github.com/michaelway/ggkm
}
\examples{
ggkm(survfit(Surv(RFSMonths, RFSStatus) ~ riskGroup), timeby = 12)

}
