% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringmanipulation.R
\name{intersectR}
\alias{intersectR}
\title{A recursive function getting intersection from mutiple vectors.
does not promise to keep any sorting}
\usage{
intersectR(...)
}
\arguments{
\item{...}{Two or more vectors}
}
\value{
a vector with overlapping values from all passed vectors
}
\description{
A recursive function getting intersection from mutiple vectors.
does not promise to keep any sorting
}
\examples{
v1 = c(1,2,3,4,5)
v2 = c(1,2,3,4)
v3 = c(2,3,4,5)
v4 = c(4,3,2,6,1)

intersectR(v1, v2, v3, v4)
#returns 4, 3, 2

}
