% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{multiplot}
\alias{multiplot}
\title{A function for plotting mutiple gg plots
Found at http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)}

\item{plotlist}{ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)}

\item{cols}{Number of columns in layout}

\item{layout}{A matrix specifying the layout. If present, 'cols' is ignored}
}
\value{
nothing
}
\description{
ggplot objects can be passed in ..., or to plotlist (as a list of ggplot objects)
}
\examples{

If the layout is something like matrix(c(1,2,3,3), nrow=2, byrow=TRUE),
}
