% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringmanipulation.R
\name{patientIDFromSample}
\alias{patientIDFromSample}
\title{A function for getting patient ID from biobank labels}
\usage{
patientIDFromSample(sampleNames, asNum = TRUE)
}
\arguments{
\item{asNum}{Should the results be return as integers, i.e. numeric. Defaults to TRUE}

\item{samplesName}{One or more biobank labels. String pattern: XNNXNNNNNX, where X = letter, N = number, e.g. P28B00005E}
}
\value{
a vector with patient ids
}
\description{
A function for getting patient ID from biobank labels
}
