% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfmanipulation.R
\name{sortby}
\alias{sortby}
\title{A function for sorting a data.frame by colname}
\usage{
sortby(x, by, decreasing = FALSE, na.last = NA)
}
\arguments{
\item{x}{data.frame}

\item{by}{colname (string) to sort by}

\item{decreasing}{passed on, defaults to FALSE}

\item{na.last}{passed on, defaults to NA}
}
\value{
A sorted data.frame
}
\description{
A function for sorting a data.frame by colname
}
\examples{
set.seed(42)
df = data.frame(start = 1:20,
                ran = sample(1:20, replace = TRUE))

sortby(df, "ran")
}
