% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stringmanipulation.R
\name{valueConvertGrep}
\alias{valueConvertGrep}
\title{A function replacing mutiple values from a vector, but using grep.
Thus, patterns can be part of the values, or any other pattern grep understands.}
\usage{
valueConvertGrep(valueData, patterns, replacements, other = NA)
}
\arguments{
\item{valueData}{Input}

\item{patterns}{A vector containing which values should be replaced}

\item{replacements}{A vector containing replacement values}

\item{other}{Any vales which is not found in 'patterns' is set to this. Defaults to NA}
}
\value{
a vector with the replaced values

#' @example

origin = c("gleason 3+3", "gleason 3+3=6", "gleason 3+3", "gleason 4+3=7", "gleason 3+4=7", "gleason 3+3", "gleason 7")
patternVector = c("3[+]3", "4[+]3", "3[+]4")
replacementVector = c("1", "3", "2")
valueConvert(origin, patternVector, replacementVector, "unknown")

#returns "1", "1", "1", "3", "2", "1", "unknown"
}
\description{
A function replacing mutiple values from a vector, but using grep.
Thus, patterns can be part of the values, or any other pattern grep understands.
}
